Rename and refactor Storage.java
